+++ src/components/leftTabWindow3.vue
@@ -1,5 +1,27 @@
 <template>
   <v-card flat class="leftPanelCard">
-    <v-card-text> </v-card-text>
+    <v-card-text>
+      <v-btn block flat @click="callAll" text="Editor Manager Output" />
+    </v-card-text>
   </v-card>
 </template>
+
+<script setup>
+import { editorManager } from '../js/editorManager'
+
+const callAll = () => {
+  if (editorManager.getEditorCount() === 0) {
+    console.log('No editors found.')
+    return
+  }
+
+  editorManager.getAllEditorIds().forEach((id) => {
+    console.log('Editor ID:', id)
+  })
+
+  console.log('--------------')
+  console.log(editorManager.getAllEditorsAsArray())
+  console.log('--------------')
+  console.log('Editor Count:', editorManager.getEditorCount())
+}
+</script>
